+++ packages/stacked/example/lib/ui/bottom_nav/history/history_viewmodel.dart
@@ -1,7 +1,5 @@
-import 'package:injectable/injectable.dart';
 import 'package:stacked/stacked.dart';
 
-@singleton
 class HistoryViewModel extends FutureViewModel<int> {
   @override
   Future<int> futureToRun() async {