+++ packages/stacked_localisation/stacked_localisation/example/lib/main.dart
@@ -1,4 +1,4 @@
-import 'package:flutter/material.dart';
+import 'package:flutter/material.dart' hide Router;
 import 'package:stacked_localisation/stacked_localisation.dart';
 import 'package:stacked_services/stacked_services.dart';