+++ packages/stacked_services/CHANGELOG.md
@@ -1,3 +1,45 @@
+## 0.7.3
+
+- Removes the warning from the get package
+- Complies with "Replaced reference to obsolete FlatButton button class in SnackBar"
+- Support `ChangeNotifier` in `ReactiveServiceMixin`
+
+## 0.7.2
+
+- Fixes a regression bug for allowing null title and content for a dialog
+
+## 0.7.1+3
+
+- Fixes onTap signature
+
+## 0.7.1+2
+
+- Fixes snackBar config null exception when using builder
+
+## 0.7.1+1
+
+- Removes requirement for config
+
+## 0.7.1
+
+- Adds `snackbarConfigBuilder` to allow for building the config at runtime instead of before `runApp` is called
+
+## 0.7.0
+
+- Adds custom mainButtonBuilder to allow you to build a main button UI from the variant passed in
+
+## 0.6.8
+
+- Adds isOpen functionality
+
+## 0.6.7
+
+- Made dialog title or content optional
+
+## 0.6.6
+
+- Adds title and message textAlign to the config
+
 ## 0.6.5
 
 - Allows for null title and message for custom snackbar